playing across multiple tabs "Works"
You need to click before the move is implemented and there are a couple things such as the timer being out of sink and the fact that when you start the game the timers start stright away and if white moves before black joins the program freezes to fix this i pla to add a simple puase screen at the start and i want to host the game timer on the server instead that way the times are the same on both screen's. TO fix the clicking i just need to move some of the board updating methods out of the pointer down function.